requirements.txt
fastapi
uvicorn
pydantic
pandas
google-generativeai
anthropic
pytest
